+++ tests/test_rloo_trainer.py
@@ -22,442 +22,13 @@ from transformers import AutoModelForCausalLM, AutoModelForSequenceClassificatio
 from transformers.testing_utils import require_peft
 from transformers.utils import is_peft_available
 
-from trl.trainer.utils import shuffle_sequence_dict
 from trl import RLOOConfig, RLOOTrainer
-from trl.trainer.rloo_trainer import (
-    RepeatSampler,
-    split_tensor_dict,
-    truncate_with_protected_tokens,
-)
 
 from .testing_utils import TrlTestCase, require_vllm
 
 
 if is_peft_available():
     from peft import LoraConfig, PeftModel
-
-
-class SplitTensorDictTester(TrlTestCase):
-    def test_split_equal_chunks(self):
-        x = torch.arange(12).reshape(6, 2)
-        y = torch.arange(6).reshape(6, 1)
-        tensor_dict = {"x": x, "y": y}
-
-        result = split_tensor_dict(tensor_dict, 3)
-
-        expected_x_chunks = torch.chunk(x, 3, dim=0)
-        expected_y_chunks = torch.chunk(y, 3, dim=0)
-        self.assertEqual(len(result), 3)
-        for i in range(3):
-            self.assertTrue(torch.equal(result[i]["x"], expected_x_chunks[i]))
-            self.assertTrue(torch.equal(result[i]["y"], expected_y_chunks[i]))
-
-    def test_with_none_tensor(self):
-        x = torch.arange(12).reshape(6, 2)
-        tensor_dict = {"x": x, "y": None}
-
-        result = split_tensor_dict(tensor_dict, 2)
-
-        expected_x_chunks = torch.chunk(x, 2, dim=0)
-        self.assertEqual(len(result), 2)
-        for i in range(2):
-            self.assertTrue(torch.equal(result[i]["x"], expected_x_chunks[i]))
-            self.assertIsNone(result[i]["y"])
-
-    def test_with_scalar(self):
-        x = torch.arange(12).reshape(6, 2)
-        tensor_dict = {"x": x, "y": torch.tensor(1)}
-
-        result = split_tensor_dict(tensor_dict, 2)
-
-        expected_x_chunks = torch.chunk(x, 2, dim=0)
-        self.assertEqual(len(result), 2)
-        for i in range(2):
-            self.assertTrue(torch.equal(result[i]["x"], expected_x_chunks[i]))
-            self.assertTrue(torch.equal(result[i]["y"], torch.tensor(1)))
-
-
-class ShuffleSequenceDictTester(TrlTestCase):
-    def test_shuffle_preserves_shape(self):
-        x = torch.arange(6).reshape(3, 2)
-        y = torch.arange(3).reshape(3, 1)
-        tensor_dict = {"x": x.clone(), "y": y.clone()}
-
-        shuffled = shuffle_sequence_dict(tensor_dict)
-
-        self.assertEqual(shuffled["x"].shape, x.shape)
-        self.assertEqual(shuffled["y"].shape, y.shape)
-
-    def test_shuffle_consistent_across_tensors(self):
-        # Use known patterns to check alignment
-        x = torch.tensor([[10, 11], [20, 21], [30, 31]])
-        y = torch.tensor([[1], [2], [3]])
-        tensor_dict = {"x": x.clone(), "y": y.clone()}
-
-        shuffled = shuffle_sequence_dict(tensor_dict)
-
-        # Build a reverse map from shuffled x rows to y values
-        for i in range(3):
-            x_row = shuffled["x"][i]
-            y_val = shuffled["y"][i].item()
-
-            if torch.equal(x_row, torch.tensor([10, 11])):
-                self.assertEqual(y_val, 1)
-            elif torch.equal(x_row, torch.tensor([20, 21])):
-                self.assertEqual(y_val, 2)
-            elif torch.equal(x_row, torch.tensor([30, 31])):
-                self.assertEqual(y_val, 3)
-            else:
-                self.fail("Unexpected x row in shuffled output.")
-
-    def test_none_tensor_remains_none(self):
-        x = torch.arange(6).reshape(3, 2)
-        tensor_dict = {"x": x.clone(), "y": None}
-
-        shuffled = shuffle_sequence_dict(tensor_dict)
-
-        self.assertIsNone(shuffled["y"])
-        self.assertEqual(shuffled["x"].shape, x.shape)
-
-    def test_shuffle_with_list(self):
-        x = torch.tensor([[10, 11], [20, 21], [30, 31]])
-        y = ["a", "b", "c"]
-
-        sequence_dict = {"x": x.clone(), "y": y}
-
-        shuffled = shuffle_sequence_dict(sequence_dict)
-
-        # Check that the list y is shuffled in the same order as x
-        for i in range(3):
-            x_row = shuffled["x"][i]
-            y_val = shuffled["y"][i]
-
-            if torch.equal(x_row, torch.tensor([10, 11])):
-                self.assertEqual(y_val, "a")
-            elif torch.equal(x_row, torch.tensor([20, 21])):
-                self.assertEqual(y_val, "b")
-            elif torch.equal(x_row, torch.tensor([30, 31])):
-                self.assertEqual(y_val, "c")
-            else:
-                self.fail("Unexpected x row in shuffled output.")
-
-
-class RepeatRandomSamplerTester(TrlTestCase):
-    def test_sampler(self):
-        dataset = ["a", "b", "c", "d", "e", "f", "g"]
-        sampler = RepeatSampler(dataset, mini_repeat_count=2)
-        # Should output something like [4, 4, 3, 3, 0, 0, 1, 1, 2, 2, 6, 6, 5, 5]
-        sampled = list(sampler)
-        # Check that the length is doubled
-        assert len(sampled) == 2 * len(dataset)
-        # Check that all indexes are present
-        assert set(sampled) == set(range(len(dataset)))
-        # Check that each element is repeated twice
-        assert all(sampled[i] == sampled[i + 1] for i in range(0, len(sampled), 2))
-
-    def test_sampler_no_shuffle(self):
-        dataset = ["a", "b", "c", "d", "e", "f", "g"]
-        sampler = RepeatSampler(dataset, mini_repeat_count=2, shuffle=False)
-        sampled = list(sampler)
-        expected = [0, 0, 1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6]
-        self.assertEqual(sampled, expected)
-
-    def test_sampler_no_repeat(self):
-        dataset = ["a", "b", "c", "d", "e", "f", "g"]
-        sampler = RepeatSampler(dataset, mini_repeat_count=1)
-        # Should output something like [4, 3, 0, 1, 2, 6, 5]
-        sampled = list(sampler)
-        # Check that the length is the same
-        assert len(sampled) == len(dataset)
-        # Check that all indexes are present
-        assert set(sampled) == set(range(len(dataset)))
-
-    def test_sampler_with_batch_size(self):
-        dataset = ["a", "b", "c", "d", "e", "f", "g", "h"]
-        sampler = RepeatSampler(dataset, mini_repeat_count=1, batch_size=2, repeat_count=2)
-        # Should output something like [4, 3, 4, 3, 0, 1, 0, 1, 2, 6, 2, 6, 5, 7, 5, 7]
-        sampled = list(sampler)
-        # Check that the length is doubled
-        assert len(sampled) == 2 * len(dataset)
-        # Check that all indexes are present
-        assert set(sampled) == set(range(len(dataset)))
-        # Check that each element is repeated as expected
-        assert all(sampled[i : i + 1] == sampled[i + 2 : i + 3] for i in range(0, len(sampled), 4))
-
-    def test_sampler_with_batch_size_and_drop(self):
-        dataset = ["a", "b", "c", "d", "e", "f", "g"]
-        sampler = RepeatSampler(dataset, mini_repeat_count=1, batch_size=2, repeat_count=2)
-        # Should output something like [4, 3, 4, 3, 0, 1, 0, 1, 2, 6, 2, 6]
-        sampled = list(sampler)
-        # Check that the length is doubled
-        assert len(sampled) == 2 * (
-            len(dataset) - 1
-        )  # one element is dropped, because it's not enough to form a batch
-        assert len(sampler) == len(sampled)  # the length should be the same as the sampled length
-        # Check that the sampled indexes are a subset of the dataset indexes
-        assert set(sampled).issubset(set(range(len(dataset))))
-        # Check that each element is repeated as expected
-        assert all(sampled[i : i + 1] == sampled[i + 2 : i + 3] for i in range(0, len(sampled), 4))
-
-    def test_sampler_with_mini_repeat_count_and_batch_size_1(self):
-        dataset = ["a", "b", "c", "d", "e", "f", "g"]
-        sampler = RepeatSampler(dataset, mini_repeat_count=2, batch_size=3, repeat_count=2)
-        # Should output something like [4, 4, 3, 3, 0, 0, 4, 4, 3, 3, 0, 0,
-        #                               1, 1, 2, 2, 6, 6, 1, 1, 2, 2, 6, 6]
-        sampled = list(sampler)
-        # Check that the length is quadrupled
-        assert len(sampled) == 4 * (len(dataset) - 1)  # 1 element is dropped, because it's not enough to form a batch
-        assert len(sampler) == len(sampled)  # the length should be the same as the sampled length
-        # Check that the sampled indexes are a subset of the dataset indexes
-        assert set(sampled).issubset(set(range(len(dataset))))
-        # Check that each element is repeated as expected
-        assert all(sampled[i] == sampled[i + 1] for i in range(0, len(sampled), 2))
-        # Check that the batch is repeated as expected
-        assert sampled[0:6] == sampled[6:12]
-        assert sampled[12:18] == sampled[18:24]
-
-    def test_sampler_with_mini_repeat_count_and_batch_size_2(self):
-        dataset = ["a", "b", "c", "d", "e", "f", "g"]
-        sampler = RepeatSampler(dataset, mini_repeat_count=3, batch_size=2, repeat_count=2)
-        # Should output something like [4, 4, 4, 3, 3, 3, 4, 4, 4, 3, 3, 3,
-        #                               0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1,
-        #                               2, 2, 2, 6, 6, 6, 2, 2, 2, 6, 6, 6]
-        sampled = list(sampler)
-        # Check that the length is sextupled
-        assert len(sampled) == 6 * (len(dataset) - 1)  # 1 element is dropped, because it's not enough to form a batch
-        assert len(sampler) == len(sampled)  # the length should be the same as the sampled length
-        # Check that the sampled indexes are a subset of the dataset indexes
-        assert set(sampled).issubset(set(range(len(dataset))))
-        # Check that each element is repeated as expected
-        assert all(sampled[i] == sampled[i + 1] == sampled[i + 2] for i in range(0, len(sampled), 3))
-        # Check that the batch is repeated as expected
-        assert sampled[0:6] == sampled[6:12]
-        assert sampled[12:18] == sampled[18:24]
-        assert sampled[24:30] == sampled[30:36]
-
-    def test_sampler_with_mini_repeat_count_and_batch_size_3(self):
-        dataset = ["a", "b", "c", "d", "e", "f", "g"]
-        sampler = RepeatSampler(dataset, mini_repeat_count=2, batch_size=2, repeat_count=3)
-        # Should output something like [4, 4, 3, 3, 4, 4, 3, 3, 4, 4, 3, 3,
-        #                               0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1,
-        #                               2, 2, 6, 6, 2, 2, 6, 6, 2, 2, 6, 6]
-        sampled = list(sampler)
-        # Check that the length is sextupled
-        assert len(sampled) == 6 * (len(dataset) - 1)  # 1 element is dropped, because it's not enough to form a batch
-        # Check that the sampled indexes are a subset of the dataset indexes
-        assert set(sampled).issubset(set(range(len(dataset))))
-        # Check that each element is repeated as expected
-        assert all(sampled[i] == sampled[i + 1] for i in range(0, len(sampled), 2))
-        # Check that the batch is repeated as expected
-        assert sampled[0:4] == sampled[4:8] == sampled[8:12]
-        assert sampled[12:16] == sampled[16:20] == sampled[20:24]
-        assert sampled[24:28] == sampled[28:32] == sampled[32:36]
-
-
-class TruncateWithProtectedTokensTester(TrlTestCase):
-    def test_basic_example(self):
-        """Test the basic example from the problem description."""
-        prompt_ids = torch.tensor([[1, 2, 3, 4, 5], [6, 7, 8, 9, 10]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = [2, 3, 6]
-        target_length = 3
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        expected_ids = torch.tensor([[2, 3, 5], [6, 9, 10]])
-        expected_mask = torch.ones_like(expected_ids)
-
-        self.assertTrue(torch.equal(new_ids, expected_ids))
-        self.assertTrue(torch.equal(new_mask, expected_mask))
-
-    def test_no_truncation_needed(self):
-        """Test when target length equals current length."""
-        prompt_ids = torch.tensor([[1, 2, 3]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = [2]
-        target_length = 3
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        self.assertTrue(torch.equal(new_ids, prompt_ids))
-        self.assertTrue(torch.equal(new_mask, prompt_mask))
-
-    def test_no_protected_tokens(self):
-        """Test truncation with no protected tokens (normal right truncation)."""
-        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = []
-        target_length = 3
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        expected_ids = torch.tensor([[3, 4, 5]])  # Last 3 tokens
-        self.assertTrue(torch.equal(new_ids, expected_ids))
-
-    def test_all_tokens_protected(self):
-        """Test when all remaining tokens are protected."""
-        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = [3, 4, 5]
-        target_length = 3
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        expected_ids = torch.tensor([[3, 4, 5]])
-        self.assertTrue(torch.equal(new_ids, expected_ids))
-
-    def test_too_many_protected_tokens(self):
-        """Test error when too many protected tokens for target length."""
-        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = [1, 2, 3, 4]
-        target_length = 3
-
-        with self.assertRaises(ValueError):
-            truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-    def test_single_batch_single_token(self):
-        """Test edge case with single batch and single token."""
-        prompt_ids = torch.tensor([[5]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = [5]
-        target_length = 1
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        self.assertTrue(torch.equal(new_ids, prompt_ids))
-
-    def test_mask_preservation(self):
-        """Test that mask values are correctly preserved."""
-        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
-        prompt_mask = torch.tensor([[1, 0, 1, 0, 1]])  # Mixed mask values
-        protected_tokens = [2, 4]
-        target_length = 3
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        expected_ids = torch.tensor([[2, 4, 5]])
-        expected_mask = torch.tensor([[0, 0, 1]])  # Corresponding mask values
-
-        self.assertTrue(torch.equal(new_ids, expected_ids))
-        self.assertTrue(torch.equal(new_mask, expected_mask))
-
-    def test_multiple_batches_different_protected(self):
-        """Test multiple batches where protected tokens appear differently."""
-        prompt_ids = torch.tensor([[1, 2, 3, 4, 5], [2, 6, 7, 8, 9], [10, 11, 12, 2, 13]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = [2]
-        target_length = 3
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        expected_ids = torch.tensor(
-            [
-                [2, 4, 5],  # 2 is protected, keep last 2 non-protected (4,5)
-                [2, 8, 9],  # 2 is protected, keep last 2 non-protected (8,9)
-                [12, 2, 13],  # 2 is protected, keep last 2 non-protected (12,13)
-            ]
-        )
-
-        self.assertTrue(torch.equal(new_ids, expected_ids))
-
-    def test_order_preservation(self):
-        """Test that relative order is preserved."""
-        prompt_ids = torch.tensor([[10, 2, 20, 3, 30, 40]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = [2, 3]
-        target_length = 4
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        # Should keep protected tokens 2,3 and last 2 non-protected tokens 30,40
-        # Order should be: 2, 3, 30, 40 (maintaining original relative positions)
-        expected_ids = torch.tensor([[2, 3, 30, 40]])
-
-        self.assertTrue(torch.equal(new_ids, expected_ids))
-
-    def test_empty_protected_tokens_list(self):
-        """Test with empty protected tokens list."""
-        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
-        prompt_mask = torch.ones_like(prompt_ids)
-        protected_tokens = []
-        target_length = 2
-
-        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
-
-        expected_ids = torch.tensor([[4, 5]])  # Last 2 tokens
-        self.assertTrue(torch.equal(new_ids, expected_ids))
-
-
-class GetHighEntropyMaskTester(TrlTestCase):
-    def get_high_entropy_mask(self, entropies, mask, threshold):
-        """Helper method to test the get_high_entropy_mask functionality."""
-        # Create a mock trainer with minimal setup
-        from unittest.mock import Mock
-
-        # Create a mock accelerator
-        mock_accelerator = Mock()
-        mock_accelerator.num_processes = 1  # Single process for testing
-
-        # Create a minimal trainer instance just to access the method
-        trainer = Mock(spec=RLOOTrainer)
-        trainer.accelerator = mock_accelerator
-        trainer.accelerator.gather = lambda x: x  # Mock gather to return the input directly
-
-        # Call the actual method from RLOOTrainer
-        return RLOOTrainer.get_high_entropy_mask(trainer, entropies, mask, threshold)
-
-    def test_compute_entropy_mask_0(self):
-        # We have a total of 12 tokens out of which 10 are non-pad.
-        # for a top_entropy_quantile of 0.8, we expect the top 20% i.e 2 non-pad tokens corresponding to
-        # the highest entropy to be unmasked.
-        # In our example these will be the tokens corresponding to the entropies 0.9 and 1.0 since 1.1 and 1.2 are pad
-        # tokens they are excluded from the entropy threshold calculation.
-        entropies = torch.tensor([[0.1, 0.2, 0.3, 0.4, 0.5, 0.6], [0.7, 0.8, 0.9, 1.0, 1.1, 1.2]])
-        mask = torch.tensor([[1, 1, 1, 1, 1, 1], [1, 1, 1, 1, 0, 0]])
-        entropy_mask = self.get_high_entropy_mask(entropies, mask, threshold=0.8)
-        expected_mask = torch.tensor([[0, 0, 0, 0, 0, 0], [0, 0, 1, 1, 0, 0]], dtype=torch.bool)
-        torch.testing.assert_close(entropy_mask, expected_mask)
-
-    def test_compute_entropy_mask_1(self):
-        # Another example with a different set of entropies and a different mask.
-        entropies = torch.tensor([[0.1, 0.2, 0.3, 1.4, 0.5, 0.14], [0.5, 0.6, 0.7, 0.8, 0.9, 1.0]])
-        mask = torch.tensor([[1, 1, 1, 1, 0, 0], [1, 1, 1, 1, 0, 0]])
-        entropy_mask = self.get_high_entropy_mask(entropies, mask, threshold=0.8)
-        expected_mask = torch.tensor([[0, 0, 0, 1, 0, 0], [0, 0, 0, 1, 0, 0]], dtype=torch.bool)
-        torch.testing.assert_close(entropy_mask, expected_mask)
-
-    def test_compute_entropy_mask_lower_threshold(self):
-        # For a threshold of 0.5 we expect the top half of the non-pad tokens to be unmasked.
-        entropies = torch.tensor([[0.1, 0.2, 0.3, 0.4, 0.5, 0.6], [0.7, 0.8, 0.9, 1.0, 1.1, 1.2]])
-        mask = torch.tensor([[1, 1, 1, 1, 1, 1], [1, 1, 1, 1, 0, 0]])
-        entropy_mask = self.get_high_entropy_mask(entropies, mask, threshold=0.5)
-        expected_mask = torch.tensor([[0, 0, 0, 0, 0, 1], [1, 1, 1, 1, 0, 0]], dtype=torch.bool)
-        torch.testing.assert_close(entropy_mask, expected_mask)
-
-    def test_compute_entropy_threshold_0(self):
-        # If the threshold is 0.0 then we expect the mask to be all ones for non-pad tokens.
-        entropies = torch.tensor([[0.1, 0.2, 0.3, 0.4, 0.5, 0.6], [0.7, 0.8, 0.9, 1.0, 1.1, 1.2]])
-        mask = torch.tensor([[1, 1, 1, 1, 1, 1], [1, 1, 1, 1, 0, 0]])
-        entropy_mask = self.get_high_entropy_mask(entropies, mask, threshold=0.0)
-        expected_mask = torch.tensor([[1, 1, 1, 1, 1, 1], [1, 1, 1, 1, 0, 0]], dtype=torch.bool)
-        torch.testing.assert_close(entropy_mask, expected_mask)
-
-    def test_compute_entropy_threshold_1(self):
-        # If the threshold is 1.0 then we expect the mask to be all zeros BUT ONE VALUE.
-        entropies = torch.tensor([[0.1, 0.2, 0.3, 0.4, 0.5, 0.6], [0.7, 0.8, 0.9, 1.0, 1.1, 1.2]])
-        mask = torch.tensor([[1, 1, 1, 1, 1, 1], [1, 1, 1, 1, 0, 0]])
-        entropy_mask = self.get_high_entropy_mask(entropies, mask, threshold=1.0)
-        expected_mask = torch.tensor([[0, 0, 0, 0, 0, 0], [0, 0, 0, 1, 0, 0]], dtype=torch.bool)
-        torch.testing.assert_close(entropy_mask, expected_mask)
-
-    def test_compute_entropy_all_masked(self):
-        # If there are no non-pad tokens we expect the mask to be all zeros.
-        entropies = torch.tensor([[0.1, 0.2, 0.3, 0.4, 0.5, 0.6], [0.7, 0.8, 0.9, 1.0, 1.1, 1.2]])
-        mask = torch.tensor([[0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0]])
-        entropy_mask = self.get_high_entropy_mask(entropies, mask, threshold=0.5)
-        expected_mask = torch.tensor([[0, 0, 0, 0, 0, 0], [0, 0, 0, 0, 0, 0]], dtype=torch.bool)
-        torch.testing.assert_close(entropy_mask, expected_mask)
 
 
 class RLOOTrainerTester(TrlTestCase):
