+++ tests/test_utils.py
@@ -22,12 +22,13 @@ from datasets import load_dataset
 from parameterized import parameterized
 from transformers import AutoModelForCausalLM, AutoTokenizer, GenerationConfig
 from transformers.testing_utils import require_peft
-from transformers.utils import is_peft_available
+from transformers.utils import is_peft_available, require_rich
 
 from trl import ModelConfig
 from trl.trainer import compute_accuracy
 from trl.trainer.utils import (
     DataCollatorForChatML,
+    RepeatSampler,
     batch_generation,
     decode_and_strip_padding,
     entropy_from_logits,
@@ -38,9 +39,14 @@ from trl.trainer.utils import (
     pad,
     print_prompt_completions_sample,
     selective_log_softmax,
+    shuffle_sequence_dict,
+    split_pixel_values_by_grid,
+    split_tensor_dict,
+    truncate_with_protected_tokens,
+    unsplit_pixel_values_by_grid,
 )
 
-from .testing_utils import TrlTestCase, require_rich
+from .testing_utils import TrlTestCase
 
 
 if is_peft_available():
@@ -539,25 +545,137 @@ class TestFlushRight(TrlTestCase):
         self.assertTrue(torch.equal(new_mask, expected_mask))
 
 
-class TestSelectiveLogSoftmax(TrlTestCase):
-    @parameterized.expand([(torch.float64,), (torch.float32,), (torch.float16,), (torch.bfloat16,)])
-    def test_selective_log_softmax(self, dtype):
-        """Test selective_log_softmax with logits of different dtypes"""
-        vocab_size = 1024
-        batch_size = 4
-        seq_len = 32
+class RepeatRandomSamplerTester(TrlTestCase):
+    def test_sampler(self):
+        dataset = ["a", "b", "c", "d", "e", "f", "g"]
+        sampler = RepeatSampler(dataset, mini_repeat_count=2)
+        # Should output something like [4, 4, 3, 3, 0, 0, 1, 1, 2, 2, 6, 6, 5, 5]
+        sampled = list(sampler)
+        # Check that the length is doubled
+        assert len(sampled) == 2 * len(dataset)
+        # Check that all indexes are present
+        assert set(sampled) == set(range(len(dataset)))
+        # Check that each element is repeated twice
+        assert all(sampled[i] == sampled[i + 1] for i in range(0, len(sampled), 2))
 
-        input_ids = torch.randint(low=0, high=vocab_size, size=(batch_size, seq_len))
+    def test_sampler_no_shuffle(self):
+        dataset = ["a", "b", "c", "d", "e", "f", "g"]
+        sampler = RepeatSampler(dataset, mini_repeat_count=2, shuffle=False)
+        sampled = list(sampler)
+        expected = [0, 0, 1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6]
+        self.assertEqual(sampled, expected)
+
+    def test_sampler_no_repeat(self):
+        dataset = ["a", "b", "c", "d", "e", "f", "g"]
+        sampler = RepeatSampler(dataset, mini_repeat_count=1)
+        # Should output something like [4, 3, 0, 1, 2, 6, 5]
+        sampled = list(sampler)
+        # Check that the length is the same
+        assert len(sampled) == len(dataset)
+        # Check that all indexes are present
+        assert set(sampled) == set(range(len(dataset)))
+
+    def test_sampler_with_batch_size(self):
+        dataset = ["a", "b", "c", "d", "e", "f", "g", "h"]
+        sampler = RepeatSampler(dataset, mini_repeat_count=1, batch_size=2, repeat_count=2)
+        # Should output something like [4, 3, 4, 3, 0, 1, 0, 1, 2, 6, 2, 6, 5, 7, 5, 7]
+        sampled = list(sampler)
+        # Check that the length is doubled
+        assert len(sampled) == 2 * len(dataset)
+        # Check that all indexes are present
+        assert set(sampled) == set(range(len(dataset)))
+        # Check that each element is repeated as expected
+        assert all(sampled[i : i + 1] == sampled[i + 2 : i + 3] for i in range(0, len(sampled), 4))
+
+    def test_sampler_with_batch_size_and_drop(self):
+        dataset = ["a", "b", "c", "d", "e", "f", "g"]
+        sampler = RepeatSampler(dataset, mini_repeat_count=1, batch_size=2, repeat_count=2)
+        # Should output something like [4, 3, 4, 3, 0, 1, 0, 1, 2, 6, 2, 6]
+        sampled = list(sampler)
+        # Check that the length is doubled
+        assert len(sampled) == 2 * (
+            len(dataset) - 1
+        )  # one element is dropped, because it's not enough to form a batch
+        assert len(sampler) == len(sampled)  # the length should be the same as the sampled length
+        # Check that the sampled indexes are a subset of the dataset indexes
+        assert set(sampled).issubset(set(range(len(dataset))))
+        # Check that each element is repeated as expected
+        assert all(sampled[i : i + 1] == sampled[i + 2 : i + 3] for i in range(0, len(sampled), 4))
+
+    def test_sampler_with_mini_repeat_count_and_batch_size_1(self):
+        dataset = ["a", "b", "c", "d", "e", "f", "g"]
+        sampler = RepeatSampler(dataset, mini_repeat_count=2, batch_size=3, repeat_count=2)
+        # Should output something like [4, 4, 3, 3, 0, 0, 4, 4, 3, 3, 0, 0,
+        #                               1, 1, 2, 2, 6, 6, 1, 1, 2, 2, 6, 6]
+        sampled = list(sampler)
+        # Check that the length is quadrupled
+        assert len(sampled) == 4 * (len(dataset) - 1)  # 1 element is dropped, because it's not enough to form a batch
+        assert len(sampler) == len(sampled)  # the length should be the same as the sampled length
+        # Check that the sampled indexes are a subset of the dataset indexes
+        assert set(sampled).issubset(set(range(len(dataset))))
+        # Check that each element is repeated as expected
+        assert all(sampled[i] == sampled[i + 1] for i in range(0, len(sampled), 2))
+        # Check that the batch is repeated as expected
+        assert sampled[0:6] == sampled[6:12]
+        assert sampled[12:18] == sampled[18:24]
+
+    def test_sampler_with_mini_repeat_count_and_batch_size_2(self):
+        dataset = ["a", "b", "c", "d", "e", "f", "g"]
+        sampler = RepeatSampler(dataset, mini_repeat_count=3, batch_size=2, repeat_count=2)
+        # Should output something like [4, 4, 4, 3, 3, 3, 4, 4, 4, 3, 3, 3,
+        #                               0, 0, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1,
+        #                               2, 2, 2, 6, 6, 6, 2, 2, 2, 6, 6, 6]
+        sampled = list(sampler)
+        # Check that the length is sextupled
+        assert len(sampled) == 6 * (len(dataset) - 1)  # 1 element is dropped, because it's not enough to form a batch
+        assert len(sampler) == len(sampled)  # the length should be the same as the sampled length
+        # Check that the sampled indexes are a subset of the dataset indexes
+        assert set(sampled).issubset(set(range(len(dataset))))
+        # Check that each element is repeated as expected
+        assert all(sampled[i] == sampled[i + 1] == sampled[i + 2] for i in range(0, len(sampled), 3))
+        # Check that the batch is repeated as expected
+        assert sampled[0:6] == sampled[6:12]
+        assert sampled[12:18] == sampled[18:24]
+        assert sampled[24:30] == sampled[30:36]
+
+    def test_sampler_with_mini_repeat_count_and_batch_size_3(self):
+        dataset = ["a", "b", "c", "d", "e", "f", "g"]
+        sampler = RepeatSampler(dataset, mini_repeat_count=2, batch_size=2, repeat_count=3)
+        # Should output something like [4, 4, 3, 3, 4, 4, 3, 3, 4, 4, 3, 3,
+        #                               0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1,
+        #                               2, 2, 6, 6, 2, 2, 6, 6, 2, 2, 6, 6]
+        sampled = list(sampler)
+        # Check that the length is sextupled
+        assert len(sampled) == 6 * (len(dataset) - 1)  # 1 element is dropped, because it's not enough to form a batch
+        # Check that the sampled indexes are a subset of the dataset indexes
+        assert set(sampled).issubset(set(range(len(dataset))))
+        # Check that each element is repeated as expected
+        assert all(sampled[i] == sampled[i + 1] for i in range(0, len(sampled), 2))
+        # Check that the batch is repeated as expected
+        assert sampled[0:4] == sampled[4:8] == sampled[8:12]
+        assert sampled[12:16] == sampled[16:20] == sampled[20:24]
+        assert sampled[24:28] == sampled[28:32] == sampled[32:36]
+
+
+class TestEntropyFromLogits(TrlTestCase):
+    @parameterized.expand(
+        [
+            (dtype, chunk_size)
+            for dtype in (torch.float64, torch.float32, torch.float16, torch.bfloat16)
+            for chunk_size in (1, 16)
+        ]
+    )
+    def test_entropy_from_logits(self, dtype, chunk_size):
+        batch_size, seq_len, vocab_size = 64, 384, 768
         logits = torch.randn(batch_size, seq_len, vocab_size, dtype=dtype)
-
-        expected_output = torch.gather(logits.log_softmax(-1), dim=-1, index=input_ids.unsqueeze(-1)).squeeze(-1)
-        actual_output = selective_log_softmax(logits, input_ids)
-
-        if dtype in [torch.float16, torch.bfloat16]:
-            # half-precision dtypes fall back to an exact method
-            self.assertTrue(torch.equal(actual_output, expected_output))
+        if dtype in (torch.float64, torch.float32):
+            p = logits.softmax(-1)
+            entropy = -torch.sum(p * p.log(), dim=-1)
         else:
-            torch.testing.assert_close(actual_output, expected_output, rtol=1e-5, atol=1e-5)
+            logps = logits.log_softmax(dim=-1)
+            entropy = -(torch.exp(logps) * logps).sum(-1)
+        predicted_entropy = entropy_from_logits(logits, chunk_size=chunk_size)
+        torch.testing.assert_close(predicted_entropy, entropy, rtol=1e-5, atol=1e-5)
 
 
 @require_rich
@@ -625,22 +743,316 @@ class TestPrintPromptCompletionsSample(TrlTestCase):
         self.assertIn(output, possible_outputs)
 
 
-class TestEntropyFromLogits(TrlTestCase):
-    @parameterized.expand(
-        [
-            (dtype, chunk_size)
-            for dtype in (torch.float64, torch.float32, torch.float16, torch.bfloat16)
-            for chunk_size in (1, 16)
-        ]
-    )
-    def test_entropy_from_logits(self, dtype, chunk_size):
-        batch_size, seq_len, vocab_size = 64, 384, 768
+class TestSelectiveLogSoftmax(TrlTestCase):
+    @parameterized.expand([(torch.float64,), (torch.float32,), (torch.float16,), (torch.bfloat16,)])
+    def test_selective_log_softmax(self, dtype):
+        """Test selective_log_softmax with logits of different dtypes"""
+        vocab_size = 1024
+        batch_size = 4
+        seq_len = 32
+
+        input_ids = torch.randint(low=0, high=vocab_size, size=(batch_size, seq_len))
         logits = torch.randn(batch_size, seq_len, vocab_size, dtype=dtype)
-        if dtype in (torch.float64, torch.float32):
-            p = logits.softmax(-1)
-            entropy = -torch.sum(p * p.log(), dim=-1)
+
+        expected_output = torch.gather(logits.log_softmax(-1), dim=-1, index=input_ids.unsqueeze(-1)).squeeze(-1)
+        actual_output = selective_log_softmax(logits, input_ids)
+
+        if dtype in [torch.float16, torch.bfloat16]:
+            # half-precision dtypes fall back to an exact method
+            self.assertTrue(torch.equal(actual_output, expected_output))
         else:
-            logps = logits.log_softmax(dim=-1)
-            entropy = -(torch.exp(logps) * logps).sum(-1)
-        predicted_entropy = entropy_from_logits(logits, chunk_size=chunk_size)
-        torch.testing.assert_close(predicted_entropy, entropy, rtol=1e-5, atol=1e-5)
+            torch.testing.assert_close(actual_output, expected_output, rtol=1e-5, atol=1e-5)
+
+
+class ShuffleSequenceDictTester(TrlTestCase):
+    def test_shuffle_preserves_shape(self):
+        x = torch.arange(6).reshape(3, 2)
+        y = torch.arange(3).reshape(3, 1)
+        tensor_dict = {"x": x.clone(), "y": y.clone()}
+
+        shuffled = shuffle_sequence_dict(tensor_dict)
+
+        self.assertEqual(shuffled["x"].shape, x.shape)
+        self.assertEqual(shuffled["y"].shape, y.shape)
+
+    def test_shuffle_consistent_across_tensors(self):
+        # Use known patterns to check alignment
+        x = torch.tensor([[10, 11], [20, 21], [30, 31]])
+        y = torch.tensor([[1], [2], [3]])
+        tensor_dict = {"x": x.clone(), "y": y.clone()}
+
+        shuffled = shuffle_sequence_dict(tensor_dict)
+
+        # Build a reverse map from shuffled x rows to y values
+        for i in range(3):
+            x_row = shuffled["x"][i]
+            y_val = shuffled["y"][i].item()
+
+            if torch.equal(x_row, torch.tensor([10, 11])):
+                self.assertEqual(y_val, 1)
+            elif torch.equal(x_row, torch.tensor([20, 21])):
+                self.assertEqual(y_val, 2)
+            elif torch.equal(x_row, torch.tensor([30, 31])):
+                self.assertEqual(y_val, 3)
+            else:
+                self.fail("Unexpected x row in shuffled output.")
+
+    def test_none_tensor_remains_none(self):
+        x = torch.arange(6).reshape(3, 2)
+        tensor_dict = {"x": x.clone(), "y": None}
+
+        shuffled = shuffle_sequence_dict(tensor_dict)
+
+        self.assertIsNone(shuffled["y"])
+        self.assertEqual(shuffled["x"].shape, x.shape)
+
+    def test_shuffle_with_list(self):
+        x = torch.tensor([[10, 11], [20, 21], [30, 31]])
+        y = ["a", "b", "c"]
+
+        sequence_dict = {"x": x.clone(), "y": y}
+
+        shuffled = shuffle_sequence_dict(sequence_dict)
+
+        # Check that the list y is shuffled in the same order as x
+        for i in range(3):
+            x_row = shuffled["x"][i]
+            y_val = shuffled["y"][i]
+
+            if torch.equal(x_row, torch.tensor([10, 11])):
+                self.assertEqual(y_val, "a")
+            elif torch.equal(x_row, torch.tensor([20, 21])):
+                self.assertEqual(y_val, "b")
+            elif torch.equal(x_row, torch.tensor([30, 31])):
+                self.assertEqual(y_val, "c")
+            else:
+                self.fail("Unexpected x row in shuffled output.")
+
+
+class SplitTensorDictTester(TrlTestCase):
+    def test_split_equal_chunks(self):
+        x = torch.arange(12).reshape(6, 2)
+        y = torch.arange(6).reshape(6, 1)
+        tensor_dict = {"x": x, "y": y}
+
+        result = split_tensor_dict(tensor_dict, 3)
+
+        expected_x_chunks = torch.chunk(x, 3, dim=0)
+        expected_y_chunks = torch.chunk(y, 3, dim=0)
+        self.assertEqual(len(result), 3)
+        for i in range(3):
+            self.assertTrue(torch.equal(result[i]["x"], expected_x_chunks[i]))
+            self.assertTrue(torch.equal(result[i]["y"], expected_y_chunks[i]))
+
+    def test_with_none_tensor(self):
+        x = torch.arange(12).reshape(6, 2)
+        tensor_dict = {"x": x, "y": None}
+
+        result = split_tensor_dict(tensor_dict, 2)
+
+        expected_x_chunks = torch.chunk(x, 2, dim=0)
+        self.assertEqual(len(result), 2)
+        for i in range(2):
+            self.assertTrue(torch.equal(result[i]["x"], expected_x_chunks[i]))
+            self.assertIsNone(result[i]["y"])
+
+    def test_with_scalar(self):
+        x = torch.arange(12).reshape(6, 2)
+        tensor_dict = {"x": x, "y": torch.tensor(1)}
+
+        result = split_tensor_dict(tensor_dict, 2)
+
+        expected_x_chunks = torch.chunk(x, 2, dim=0)
+        self.assertEqual(len(result), 2)
+        for i in range(2):
+            self.assertTrue(torch.equal(result[i]["x"], expected_x_chunks[i]))
+            self.assertTrue(torch.equal(result[i]["y"], torch.tensor(1)))
+
+
+class SplitPixelValuesByGridTester(TrlTestCase):
+    def test_split_correctly_0(self):
+        batch = {
+            "image_grid_thw": torch.tensor([[1, 2, 2], [1, 1, 4]]),  # Products: [4, 4]
+            "pixel_values": torch.arange(8 * 3).reshape(8, 3),  # Shape: [8, 3]
+        }
+        result = split_pixel_values_by_grid(batch)
+        self.assertIsInstance(result["pixel_values"], list)
+        self.assertEqual(len(result["pixel_values"]), 2)
+        self.assertTrue(torch.equal(result["pixel_values"][0], batch["pixel_values"][:4]))
+        self.assertTrue(torch.equal(result["pixel_values"][1], batch["pixel_values"][4:]))
+
+    def test_split_correctly_1(self):
+        batch = {
+            "image_grid_thw": torch.tensor([[1, 2, 2], [1, 2, 4]]),  # Products: [4, 8]
+            "pixel_values": torch.arange(12 * 3).reshape(12, 3),  # Shape: [12, 3]
+        }
+        result = split_pixel_values_by_grid(batch)
+        self.assertIsInstance(result["pixel_values"], list)
+        self.assertEqual(len(result["pixel_values"]), 2)
+        self.assertTrue(torch.equal(result["pixel_values"][0], batch["pixel_values"][:4]))
+        self.assertTrue(torch.equal(result["pixel_values"][1], batch["pixel_values"][4:12]))
+
+    def test_missing_keys(self):
+        batch = {"pixel_values": torch.tensor([1.0])}
+        result = split_pixel_values_by_grid(batch)
+        self.assertEqual(result, batch)
+
+    def test_mismatched_length(self):
+        batch = {
+            "image_grid_thw": torch.tensor([[2, 1, 1], [2, 1, 1]]),  # Total = 4
+            "pixel_values": torch.randn(3, 5),  # Only 3 rows
+        }
+        with self.assertRaises(ValueError):
+            split_pixel_values_by_grid(batch)
+
+
+class TruncateWithProtectedTokensTester(TrlTestCase):
+    def test_basic_example(self):
+        """Test the basic example from the problem description."""
+        prompt_ids = torch.tensor([[1, 2, 3, 4, 5], [6, 7, 8, 9, 10]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = [2, 3, 6]
+        target_length = 3
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        expected_ids = torch.tensor([[2, 3, 5], [6, 9, 10]])
+        expected_mask = torch.ones_like(expected_ids)
+
+        self.assertTrue(torch.equal(new_ids, expected_ids))
+        self.assertTrue(torch.equal(new_mask, expected_mask))
+
+    def test_no_truncation_needed(self):
+        """Test when target length equals current length."""
+        prompt_ids = torch.tensor([[1, 2, 3]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = [2]
+        target_length = 3
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        self.assertTrue(torch.equal(new_ids, prompt_ids))
+        self.assertTrue(torch.equal(new_mask, prompt_mask))
+
+    def test_no_protected_tokens(self):
+        """Test truncation with no protected tokens (normal right truncation)."""
+        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = []
+        target_length = 3
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        expected_ids = torch.tensor([[3, 4, 5]])  # Last 3 tokens
+        self.assertTrue(torch.equal(new_ids, expected_ids))
+
+    def test_all_tokens_protected(self):
+        """Test when all remaining tokens are protected."""
+        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = [3, 4, 5]
+        target_length = 3
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        expected_ids = torch.tensor([[3, 4, 5]])
+        self.assertTrue(torch.equal(new_ids, expected_ids))
+
+    def test_too_many_protected_tokens(self):
+        """Test error when too many protected tokens for target length."""
+        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = [1, 2, 3, 4]
+        target_length = 3
+
+        with self.assertRaises(ValueError):
+            truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+    def test_single_batch_single_token(self):
+        """Test edge case with single batch and single token."""
+        prompt_ids = torch.tensor([[5]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = [5]
+        target_length = 1
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        self.assertTrue(torch.equal(new_ids, prompt_ids))
+
+    def test_mask_preservation(self):
+        """Test that mask values are correctly preserved."""
+        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
+        prompt_mask = torch.tensor([[1, 0, 1, 0, 1]])  # Mixed mask values
+        protected_tokens = [2, 4]
+        target_length = 3
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        expected_ids = torch.tensor([[2, 4, 5]])
+        expected_mask = torch.tensor([[0, 0, 1]])  # Corresponding mask values
+
+        self.assertTrue(torch.equal(new_ids, expected_ids))
+        self.assertTrue(torch.equal(new_mask, expected_mask))
+
+    def test_multiple_batches_different_protected(self):
+        """Test multiple batches where protected tokens appear differently."""
+        prompt_ids = torch.tensor([[1, 2, 3, 4, 5], [2, 6, 7, 8, 9], [10, 11, 12, 2, 13]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = [2]
+        target_length = 3
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        expected_ids = torch.tensor(
+            [
+                [2, 4, 5],  # 2 is protected, keep last 2 non-protected (4,5)
+                [2, 8, 9],  # 2 is protected, keep last 2 non-protected (8,9)
+                [12, 2, 13],  # 2 is protected, keep last 2 non-protected (12,13)
+            ]
+        )
+
+        self.assertTrue(torch.equal(new_ids, expected_ids))
+
+    def test_order_preservation(self):
+        """Test that relative order is preserved."""
+        prompt_ids = torch.tensor([[10, 2, 20, 3, 30, 40]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = [2, 3]
+        target_length = 4
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        # Should keep protected tokens 2,3 and last 2 non-protected tokens 30,40
+        # Order should be: 2, 3, 30, 40 (maintaining original relative positions)
+        expected_ids = torch.tensor([[2, 3, 30, 40]])
+
+        self.assertTrue(torch.equal(new_ids, expected_ids))
+
+    def test_empty_protected_tokens_list(self):
+        """Test with empty protected tokens list."""
+        prompt_ids = torch.tensor([[1, 2, 3, 4, 5]])
+        prompt_mask = torch.ones_like(prompt_ids)
+        protected_tokens = []
+        target_length = 2
+
+        new_ids, new_mask = truncate_with_protected_tokens(prompt_ids, prompt_mask, target_length, protected_tokens)
+
+        expected_ids = torch.tensor([[4, 5]])  # Last 2 tokens
+        self.assertTrue(torch.equal(new_ids, expected_ids))
+
+
+class UnsplitPixelValuesByGridTester(TrlTestCase):
+    def test_unsplit_correctly(self):
+        split = [torch.randn(4, 5), torch.randn(2, 5)]
+        merged = torch.cat(split, dim=0)
+        batch = {"pixel_values": split, "other_key": torch.tensor([1])}
+        result = unsplit_pixel_values_by_grid(batch)
+        self.assertIsInstance(result["pixel_values"], torch.Tensor)
+        self.assertTrue(torch.allclose(result["pixel_values"], merged))
+        self.assertIn("other_key", result)
+
+    def test_no_op_if_not_list(self):
+        original = torch.randn(5, 3)
+        batch = {"pixel_values": original}
+        result = unsplit_pixel_values_by_grid(batch)
+        self.assertTrue(torch.equal(result["pixel_values"], original))
